requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-socketio>=5.3.0
